🚚(models) move Enrollment model to models/courses.py
The models/products.py file was getting too long and any enrollments
belong more to courses than the belong to products and orders.
